+++ SMTPLibrary/NSStream+SKPSMTPExtensions.m
@@ -35,7 +35,7 @@
 
 @implementation NSStream (SKPSMTPExtensions)
 
-+ (void)getStreamsToHostNamed:(NSString *)hostName port:(NSInteger)port inputStream:(NSInputStream **)inputStream outputStream:(NSOutputStream **)outputStream
++ (void)getStreamsToHostNamed:(NSString *)hostName port:(int)port inputStream:(NSInputStream **)inputStream outputStream:(NSOutputStream **)outputStream
 {
     CFHostRef           host;
     CFReadStreamRef     readStream;
