+++ SMTPLibrary/SKPSMTPMessage.m
@@ -329,7 +329,7 @@ NSString *kSKPSMTPPartContentTransferEncodingKey = @"kSKPSMTPPartContentTransfer
         {
             uint8_t buf[1024];
             memset(buf, 0, sizeof(uint8_t) * 1024);
-            unsigned int len = 0;
+            NSInteger len = 0;
             len = [(NSInputStream *)stream read:buf maxLength:1024];
             if(len) 
             {
